backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
firebase-admin>=6.0.0
python-multipart
python-socketio>=5.10.0

# AI/ML
torch
stable-baselines3
numpy
pandas
scikit-learn
prophet

# Computer Vision (Mobile Camera Integration)
opencv-python>=4.8.0
ultralytics>=8.0.0
pillow>=10.0.0
imutils>=0.5.4

# Mobile Camera Server
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# WebSocket for FastAPI
websockets>=12.0
